Create a SimpleInstrument class
