+++ MDANSE_GUI/Src/MDANSE_GUI/InputWidgets/InstrumentResolutionWidget.py
@@ -25,7 +25,8 @@ from qtpy.QtCore import Slot
 from qtpy.QtGui import QDoubleValidator
 
 from MDANSE_GUI.InputWidgets.WidgetBase import WidgetBase
-from MDANSE_GUI.Widgets.ResolutionDialog import ResolutionDialog, widget_text_map
+from MDANSE_GUI.Widgets.ResolutionDialog import ResolutionDialog
+from MDANSE_GUI.Widgets.ResolutionWidget import widget_text_map
 
 
 init_parameters = {
@@ -54,7 +55,7 @@ class InstrumentResolutionWidget(WidgetBase):
         self._dialog_button.clicked.connect(self.helper_dialog)
         self._dialog_button.setEnabled(True)
         self.helper = ResolutionDialog(self._base)
-        self.helper.parameters_changed.connect(self.set_parameters_from_dialog)
+        self.helper._panel.parameters_changed.connect(self.set_parameters_from_dialog)
         self._layout.addWidget(self._type_combo, 0, 1)
         self._layout.addWidget(self._dialog_button, 0, 0)
         self._labels = []
@@ -137,7 +138,7 @@ class InstrumentResolutionWidget(WidgetBase):
                 else:
                     self._fields[index].setStyleSheet("")
                 params[key] = value
-        self.helper.update_fields((function, params))
+        self.helper._panel.update_fields((function, params))
         return (function, params)
 
     @Slot(dict)
