+++ MDANSE_GUI/Src/MDANSE_GUI/Tabs/InstrumentTab.py
@@ -118,7 +118,7 @@ if __name__ == "__main__":
 
     app = QApplication(sys.argv)
     window = QMainWindow()
-    the_tab = TrajectoryTab.standard_instance()
+    the_tab = InstrumentTab.standard_instance()
     window.setCentralWidget(the_tab._core)
     window.show()
     app.exec()
